git-svn-id: https://svn.cebitec.uni-bielefeld.de/svn/mgx@1701 2f99df6c-41e9-6383-b5e2-de53e728a4e2

+++ MGX-maven-web/src/main/java/de/cebitec/mgx/web/StatisticsBean.java
@@ -44,9 +44,9 @@ public class StatisticsBean {
     @Produces("application/x-protobuf")
     @Secure(rightsNeeded = {MGXRoles.Guest})
     public PointDTOList create(MGXLongList dto) {
-        long[] data = new long[dto.getLongCount()];
-        for (int i=0;i< dto.getLongCount(); i++) {
-            data[i] = dto.getLong(i).getValue();
+        int[] data = new int[dto.getLongCount()];
+        for (int i = 0; i < dto.getLongCount(); i++) {
+            data[i] = (int) dto.getLong(i).getValue();
         }
         AutoCloseableIterator<Point> ret = rarefaction.rarefy(data);
         return PointDTOFactory.getInstance().toDTOList(ret);
